# docs/swagger.yaml
consumes:
- application/json
definitions:
...
  github_com_etkecc_mrs_internal_model.MatrixError:
    properties:
      errcode:
        description: Matrix error code, e.g M_UNAUTHORIZED
        type: string
      error:
        description: Matrix error message
        type: string
    type: object
...
  internal_controllers.reportSubmission:
    properties:
      no_msc1929:
        type: boolean
      reason:
        type: string
      roomID:
        type: string
    type: object
...
info:
  contact: {}
  description: Matrix Rooms Search crawls the Matrix federation and serves public-room
    search. It also answers part of the Matrix federation and client-server APIs for
    the rooms it has indexed. Where our implementation diverges from the spec, that
    divergence is documented here on purpose, not hidden. matrixrooms.info is a public
    demo running MRS.
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0
  title: Matrix Rooms Search API
  version: 0.0.0
paths:
...
  /discover/msc1929/{name}:
    post:
      description: 'Fetches and validates a server''s MSC1929 support file (the contacts
        in /.well-known/matrix/support). Returns 204 when it is valid, or 400 with
        a list of the specific problems: empty file, a contact missing a role, an
        unsupported role, no contacts at all, or the deprecated ''admins'' field.
        A handy pre-flight before relying on a server''s support contacts.'
      parameters:
      - description: Server name to check
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Support file is valid
        "400":
          description: Support file is missing, unreachable, or invalid
          schema:
            items:
              $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixError'
            type: array
      summary: Check a server's MSC1929 support file
      tags:
      - discovery
  /mod/ban/{room_id}:
    get:
      description: Bans a room from the index. Yes, it is a GET that mutates state,
        we know, and no, we are not proud of it. And a User-Agent containing "bot"
        gets a 403 even with valid credentials, a scar from crawlers tripping this,
        not a feature. Both are real behavior, documented on purpose so they surprise
        you here and not in production.
      parameters:
      - description: Room ID to ban
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Confirmation message
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: User-Agent contains 'bot'
      security:
      - ModerationAuth: []
      summary: Ban a room
      tags:
      - moderation
  /mod/list:
    get:
      description: Lists banned room IDs. Append /{server_name} to filter to a single
        server. An empty list is a 204, not an empty 200. As with the rest of the
        mod group, a "bot" User-Agent gets a 403 even authenticated.
      produces:
      - application/json
      responses:
        "200":
          description: Banned room IDs
          schema:
            items:
              type: string
            type: array
        "204":
          description: No banned rooms
        "403":
          description: User-Agent contains 'bot'
      security:
      - ModerationAuth: []
      summary: List banned rooms
      tags:
      - moderation
  /mod/list-reported:
    get:
      description: Lists reported rooms as a room-ID to reason map. Append /{server_name}
        to filter to a single server. Empty is a 204. A "bot" User-Agent gets a 403
        even authenticated.
      produces:
      - application/json
      responses:
        "200":
          description: Room ID to report reason
          schema:
            additionalProperties:
              type: string
            type: object
        "204":
          description: No reported rooms
        "403":
          description: User-Agent contains 'bot'
      security:
      - ModerationAuth: []
      summary: List reported rooms
      tags:
      - moderation
  /mod/report/{room_id}:
    post:
      consumes:
      - application/json
      description: 'Report a room for moderation. Open on purpose, no auth: anyone
        can flag a room, that is the whole point. The JSON body carries `reason` (5
        chars minimum, or you get a 400) and an optional `no_msc1929` flag. This is
        the one honest, no-auth POST in the moderation set.'
      parameters:
      - description: Room ID being reported
        in: path
        name: room_id
        required: true
        type: string
      - description: Report reason (5 chars minimum) and options
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.reportSubmission'
      produces:
      - application/json
      responses:
        "202":
          description: Report accepted
        "400":
          description: Invalid room ID, or reason too short or missing
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixError'
        "500":
          description: Internal error while processing the report
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixError'
      summary: Report a room
      tags:
      - moderation
  /mod/unban/{room_id}:
    get:
      description: 'Lifts a room ban. Same shape as ban: a state-changing GET, and
        a "bot" User-Agent gets a 403 even authenticated. Real behavior, owned.'
      parameters:
      - description: Room ID to unban
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Confirmation message
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: User-Agent contains 'bot'
      security:
      - ModerationAuth: []
      summary: Unban a room
      tags:
      - moderation
  /mod/unreport/{room_id}:
    get:
      description: 'Clears reports on a room. Two warts, both owned: it is a state-changing
        GET, and a User-Agent containing "bot" gets a 403 even with valid moderation
        credentials, a scar from crawlers tripping these endpoints, not a feature.
        There is also a no-room-id form (GET /mod/unreport) that hits the same handler
        with an empty ID.'
      parameters:
      - description: Room ID to clear reports for
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Reports cleared
        "403":
          description: User-Agent contains 'bot'
      security:
      - ModerationAuth: []
      summary: Clear a room's reports
      tags:
      - moderation
...
produces:
- application/json
schemes:
- https
securityDefinitions:
  AdminAuth:
    type: basic
  CatalogAuth:
    type: basic
  DiscoveryAuth:
    type: basic
  FederationAuth:
    description: 'Matrix federation signature auth: an `Authorization: X-Matrix origin=...,key=...,sig=...`
      header carrying an ed25519 signature over the request, which we verify against
      the origin server''s published keys.'
    in: header
    name: Authorization
    type: apiKey
  MetricsAuth:
    type: basic
  ModerationAuth:
    type: basic
swagger: "2.0"